+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to Major.Minor.Patch versioning.
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+
+## [Unreleased]
+### Added
+- Support for existing dependencies to be used as parameters in behavior constructors.
 
 
 ## [1.1.1] - 2022-03-07
@@ -42,6 +47,7 @@ and this project adheres to Major.Minor.Patch versioning.
 - Split showcase app from the test project.
 
 
-[Unreleased]: https://github.com/lstertz/ContextualProgramming/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/lstertz/ContextualProgramming/compare/v1.1.1...HEAD
+[1.1.1]: https://github.com/lstertz/ContextualProgramming/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/lstertz/ContextualProgramming/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/lstertz/ContextualProgramming/releases/tag/v1.0.0